agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
# plus your existing agent package and its deps
# e.g., the package that provides `agents`, AsyncOpenAI, etc.
